Alter Gulp tasks and naming conventions for compiled scss
ADD a sample Humburger button and slidein drawer nav component with test files in the new 'test' directory.
CHANGE the main gulp styles task to allow multiple files of any name to be compiled from that directory.
ADD a new gulp task (components) to allow components to be compiled separately. This deprecates the minify task (and maybe best renamed as such)